Mongo Work & Vorkath Boss
modified all necessary calls to use Mongo instead of gdoc

changed Item price to ulong to reflect game's increase in item value past 2.14b

fixed Setup assign popup to sort bosses by name

added Vorkath boss & associated setup items

+++ Assets/Scripts/Items/Item.cs
@@ -11,11 +11,11 @@ public class Item : IComparable<Item>, IValuable
     //  Properties & fields
     public int ItemId { get; set; }
     public string ItemName { get; set; }
-    public uint Price { get; set; }
+    public ulong Price { get; set; }
 
     //  Constructors
     public Item() { }
-    public Item(int itemId, string itemName, uint price)
+    public Item(int itemId, string itemName, ulong price)
     {
         this.ItemId = itemId;
         this.ItemName = itemName;
